WIP: mode0 now start with the correct DART

+++ src/lib-dart/tagion/dart/DARTOptions.d
@@ -29,7 +29,7 @@ struct DARTOptions {
     struct Synchronize {
         ulong maxSlaves;
         ulong maxMasters;
-        ulong maxSlavePort;
+        ushort maxSlavePort;
         ushort netFromAng;
         ushort netToAng;
         ulong tick_timeout;
